Get rid of splice support
It never made much sense. No one is going to use a pipe on stdout save
for an interactive task, and in interactive tasks vmsplice is utterly
broken because we reuse gifted pages. And if we don't, they are going to
be counted towards physical memory usage even if just a few bytes are
filled.